1. Split up thread pool code. 2. Fixed missing parameter in assert() and added a comment. 3. Additional comments for thread pool code. 4. Thread renamed to Proxy, ThreadWorker renamed to Thread. 5. Renamed 'call' option to 'callable' to avoid using a reserved word.
