+++ test/index.html
@@ -19,7 +19,7 @@
     <script src="core/scheduler.js"></script>
     <script src="core/component/transform.js"></script>
     <script src="core/resource/script.js"></script>
-    <script src="core/thread.js"></script>
+    <script src="core/threading/pool.js"></script>
 
     <script src="graphics/resource/mesh.js"></script>
     <style>
