+++ test/index.html.dist
@@ -16,7 +16,7 @@
     <script src="core/scheduler.js"></script>
     <script src="core/component/transform.js"></script>
     <script src="core/resource/script.js"></script>
-    <script src="core/thread.js"></script>
+    <script src="core/threading/pool.js"></script>
     <style>
       #test-canvas {
         width: 300px;
